% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_respondent_character_table.R
\name{check_awareness}
\alias{check_awareness}
\title{check_awareness}
\usage{
check_awareness(data, config)
}
\arguments{
\item{data}{}

\item{config}{}
}
\value{

}
\description{
check_awareness
}
