% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_respondent_character_table.R
\name{check_education}
\alias{check_education}
\title{check_education}
\usage{
check_education(data, config)
}
\arguments{
\item{data}{}

\item{config}{}
}
\value{

}
\description{
check_education
}
