% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_groups_recorded.R
\name{check_groups_recorded}
\alias{check_groups_recorded}
\title{check_groups_recorded}
\usage{
check_groups_recorded(data, config)
}
\arguments{
\item{data}{respondent character table}

\item{config}{config file (for paths to validated groups recorded csv)}
}
\value{
respondent data with a column added for the groups the respondent's reported recording as listed in text field of the other option on the groups recorded question.
  range-shifting species
}
\description{
check_groups_recorded
}
