% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_n_scale_age_n_years_recording.R
\name{check_n_scale_age_n_years_recording}
\alias{check_n_scale_age_n_years_recording}
\title{check_n_scale_age_n_years_recording}
\usage{
check_n_scale_age_n_years_recording(data)
}
\arguments{
\item{data}{respondent data}
}
\value{
respondent data with a column added for mean centred and scaled years
  recording and respondent age
}
\description{
check_n_scale_age_n_years_recording
}
