% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_recorder_role.R
\name{check_recorder_role}
\alias{check_recorder_role}
\title{check_recorder_role}
\usage{
check_recorder_role(data, config)
}
\arguments{
\item{data}{respondent character table}

\item{config}{config file (for paths to validated recorder roles csv)}
}
\value{

}
\description{
check_recorder_role
}
