% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_respondent_character_table.R
\name{check_respondents_ages}
\alias{check_respondents_ages}
\title{check_respondents_ages}
\usage{
check_respondents_ages(data)
}
\arguments{
\item{data}{}
}
\value{

}
\description{
check_respondents_ages
}
