% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_survey_columns.R
\name{check_respondents_postcode}
\alias{check_respondents_postcode}
\title{check_respondents_postcode}
\usage{
check_respondents_postcode(data)
}
\arguments{
\item{data}{}
}
\value{

}
\description{
check_respondents_postcode
}
