% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_respondents_reported_gender.R
\name{check_respondents_reported_gender}
\alias{check_respondents_reported_gender}
\title{check_respondents_reported_gender}
\usage{
check_respondents_reported_gender(data)
}
\arguments{
\item{data}{respondent character table}
}
\value{
dataframe with imputed gender added (not added in practice as no responses required imputing)
}
\description{
check_respondents_reported_gender
}
