% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_survey_columns.R
\name{check_wildlife_sector}
\alias{check_wildlife_sector}
\title{check_wildlife_sector}
\usage{
check_wildlife_sector(data)
}
\arguments{
\item{data}{}
}
\value{

}
\description{
check_wildlife_sector
}
