% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_survey_data.R
\name{clean_survey_data}
\alias{clean_survey_data}
\title{clean_survey_data}
\usage{
clean_survey_data(data)
}
\arguments{
\item{data}{raw survey data}
}
\value{
cleaned survey data
}
\description{
clean_survey_data applies 5 functions consecutively: remove_pilot_responses, remove_incomplete_surveys, remove_meta_info,remove_response_times, remove_instructions.
}
\seealso{
remove_pilot_responses, remove_incomplete_survey, remove_meta_info, remove_response_times, remove_instructions
}
