% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_species_rasters.R
\name{create_species_rasters}
\alias{create_species_rasters}
\title{create_species_rasters}
\usage{
create_species_rasters(config)
}
\arguments{
\item{config}{config file (for paths to raster images)}
}
\value{
list of species rasters
}
\description{
create_species_rasters
}
