% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_respondents_postcode.R
\name{format_respondents_postcode}
\alias{format_respondents_postcode}
\title{format_respondents_postcode}
\usage{
format_respondents_postcode(data)
}
\arguments{
\item{data}{dataframe (respondent character table)}
}
\value{
dataframe (respondent character table with postcodes renamed and formatted.
}
\description{
format_respondents_postcode
}
