% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_supp_figures.R
\name{generate_supp_figures}
\alias{generate_supp_figures}
\title{generate_supp_figures}
\usage{
generate_supp_figures(
  respondent_frame,
  species_frame,
  management_model,
  management_responses_to_model,
  config
)
}
\arguments{
\item{respondent_frame}{}

\item{species_frame}{}

\item{config}{}
}
\value{
a list containing all supplementary figures
}
\description{
generate_supp_figures
}
