% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_respondent_characters.R
\name{get_respondent_characters}
\alias{get_respondent_characters}
\title{get_respondent_characters}
\usage{
get_respondent_characters(data)
}
\arguments{
\item{data}{}
}
\value{

}
\description{
get_respondent_characters
}
