% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_respondent_characters.R
\name{get_species_responses}
\alias{get_species_responses}
\title{get_species_responses}
\usage{
get_species_responses(data)
}
\arguments{
\item{data}{}
}
\value{

}
\description{
get_species_responses
}
