% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getOrdinalNumber}
\alias{getOrdinalNumber}
\title{getOrdinalNumber}
\usage{
getOrdinalNumber(num)
}
\arguments{
\item{num}{a number (day in month)}
}
\value{
character vector of day in string format with ordinal suffix
}
\description{
getOrdinalNumber
}
