% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_from_responses_other.R
\name{impute_from_responses_other}
\alias{impute_from_responses_other}
\title{impute_from_responses_other}
\usage{
impute_from_responses_other(
  data,
  config,
  quiet = TRUE,
  view_imputations = FALSE
)
}
\arguments{
\item{data}{respondent data}

\item{config}{config file needed for links to validation sheets}

\item{quiet}{#TODO}

\item{view_imputations}{#TODO}
}
\value{
respondent data with a columns added for questions where imputations
  to data from open text fields were required.
}
\description{
impute_from_responses_other
}
