% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingest_survey_data.R
\name{ingest_survey_data}
\alias{ingest_survey_data}
\title{ingest_survey_data}
\usage{
ingest_survey_data(config)
}
\arguments{
\item{config}{config file with paths to survey data}
}
\value{
raw survey data
}
\description{
ingest_survey_data
}
