% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_species_in_recorded_group.R
\name{is_species_in_recorded_group}
\alias{is_species_in_recorded_group}
\title{is_species_in_recorded_group}
\usage{
is_species_in_recorded_group(speciesdata, respondentdata)
}
\arguments{
\item{speciesdata}{cleaned species responses data}

\item{respondentdata}{cleaned respondent data}
}
\value{
respondent data with a column added for match which is positive when
  a respondent records the species group of the species they were shown in
  the survey question and negative where not.
}
\description{
is_species_in_recorded_group
}
