% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_attitude_pca.R
\name{make_attitude_clusters}
\alias{make_attitude_clusters}
\title{make_attitude_clusters}
\usage{
make_attitude_clusters(mca)
}
\arguments{
\item{mca}{}
}
\value{

}
\description{
make_attitude_clusters
}
