% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_supp_figure_10.R
\name{make_attitude_pca}
\alias{make_attitude_pca}
\title{make_attitude_pca}
\usage{
make_attitude_pca(data = management_responses_to_model)
}
\arguments{
\item{data}{}
}
\value{

}
\description{
make_attitude_pca
}
