% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_supp_fig_10c.R
\name{make_supp_fig_10c}
\alias{make_supp_fig_10c}
\title{make_supp_fig_10c}
\usage{
make_supp_fig_10c(pca_output)
}
\arguments{
\item{pca_output}{pca of management attitudes}
}
\value{
lineplot of attitudes to different management options at the respondent level (facetted by cluster)
}
\description{
make_supp_fig_10c
}
