% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_tidy_datasets.R
\name{make_tidy_datasets}
\alias{make_tidy_datasets}
\title{make_tidy_datasets}
\usage{
make_tidy_datasets(data, frame = c("respondents", "species"))
}
\arguments{
\item{data}{clean survey data}

\item{frame}{either respondents or species}
}
\value{
either the starting point for a tidy dataframe containing respondents traits, or species responses.
}
\description{
make_tidy_datasets
}
