% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_word_associations.R
\name{merge_word_associations}
\alias{merge_word_associations}
\title{merge_word_associations}
\usage{
merge_word_associations(data)
}
\arguments{
\item{data}{respondent character table}
}
\value{
dataframe (respondent character table with word association questions coalesced)
}
\description{
merge_word_associations
}
