% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_species_attitudes.R
\name{model_species_attitudes}
\alias{model_species_attitudes}
\title{model_species_attitude}
\usage{
model_species_attitudes(attitudes_to_species, respondent_data)
}
\arguments{
\item{attitudes_to_species}{attitudes to species}

\item{respondent_data}{cleaned respondent data}
}
\value{
brms model object of predicted attitudes to management
}
\description{
model_species_attitude
}
