% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivot_survey_long.R
\name{pivot_species_long}
\alias{pivot_species_long}
\title{pivot_species_long}
\usage{
pivot_species_long(data)
}
\arguments{
\item{data}{survey_species_responses}
}
\value{
dataframe (nrow =  unique(respondent) x length(species) (16))
}
\description{
pivot_species_long
}
