% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fig_1.R
\name{plot_fig_1_test}
\alias{plot_fig_1_test}
\title{plot_fig_1}
\usage{
plot_fig_1_test(speciesdata)
}
\arguments{
\item{speciesdata}{survey_species_responses}
}
\value{
figure 1 (likert lattice plot)
}
\description{
plot_fig_1
}
