% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fig_2a.R
\name{plot_fig_2a}
\alias{plot_fig_2a}
\title{plot_fig_2a}
\usage{
plot_fig_2a(model)
}
\arguments{
\item{model}{species_attitude_model}
}
\value{
figure 2(a) - a ggplot of the marginal effects of familiarity
}
\description{
plot_fig_2a
}
