% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fig_2b.R
\name{plot_fig_2b}
\alias{plot_fig_2b}
\title{plot_fig_2b}
\usage{
plot_fig_2b(model, config)
}
\arguments{
\item{model}{species attitudes model}

\item{config}{config}
}
\value{
figure 2(b) ggplot
}
\description{
plot_fig_2b
}
