% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_supp_fig_1_and_2.R
\name{plot_supp_fig_1}
\alias{plot_supp_fig_1}
\title{plot_supp_fig_1}
\usage{
plot_supp_fig_1(data, config)
}
\arguments{
\item{data}{}

\item{config}{}
}
\value{

}
\description{
plot_supp_fig_1
}
