% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_supp_figure_10.R
\name{plot_supp_fig_10}
\alias{plot_supp_fig_10}
\title{plot_supp_fig_10}
\usage{
plot_supp_fig_10(data)
}
\arguments{
\item{data}{management_responses_to_model}
}
\value{
supplementary figure 10 (outputs of clustering method)
}
\description{
plot_supp_fig_10
}
