% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_supp_fig_1_and_2.R
\name{plot_supp_fig_2}
\alias{plot_supp_fig_2}
\title{plot_supp_fig_2}
\usage{
plot_supp_fig_2(data, config)
}
\arguments{
\item{data}{Survey data (cleaned)}

\item{config}{Config file (for paths to census  data)}
}
\value{
supplementary figure 2
}
\description{
plot_supp_fig_2
}
