% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_supp_fig_3.R
\name{plot_supp_fig_3}
\alias{plot_supp_fig_3}
\title{plot_supp_fig_3}
\usage{
plot_supp_fig_3(data)
}
\arguments{
\item{data}{survey data (respondent cleaned)}
}
\value{
supplementary figure 3
}
\description{
plot_supp_fig_3
}
