% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_supp_fig_4.R
\name{plot_supp_fig_4}
\alias{plot_supp_fig_4}
\title{plot_supp_fig_4}
\usage{
plot_supp_fig_4(data = respondent_table_clean)
}
\arguments{
\item{data}{cleaned respondent data table}
}
\value{
faceted plot of respondent demographics
}
\description{
produces faceted plot of respondent demographics
}
