% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_supp_fig_5.R
\name{plot_supp_fig_5}
\alias{plot_supp_fig_5}
\title{plot_supp_fig_5}
\usage{
plot_supp_fig_5(data, config)
}
\arguments{
\item{data}{cleaned respondent data table}

\item{config}{config file for paths to spatial polygons}
}
\value{
Map of respondent locations
}
\description{
plot_supp_fig_5
}
