% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_supp_fig_6.R
\name{plot_supp_fig_6}
\alias{plot_supp_fig_6}
\title{plot_supp_fig_6}
\usage{
plot_supp_fig_6(speciesdata = survey_responses_to_model)
}
\arguments{
\item{speciesdata}{}
}
\value{

}
\description{
plot_supp_fig_6
}
