% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_supp_fig_8.R
\name{plot_supp_fig_8}
\alias{plot_supp_fig_8}
\title{Title}
\usage{
plot_supp_fig_8(data)
}
\arguments{
\item{data}{brms model object of the management multivariate model}
}
\value{
supplementary figure 8 (ggplot)
}
\description{
Title
}
