% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_supp_fig_9.R
\name{plot_supp_fig_9}
\alias{plot_supp_fig_9}
\title{plot_supp_fig_9}
\usage{
plot_supp_fig_9(model)
}
\arguments{
\item{model}{Management attitudes brms model}
}
\value{
supplementary figure 9
}
\description{
plot_supp_fig_9
}
