% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_marginal_seen_and_match.R
\name{predict_marginal_seen_and_match}
\alias{predict_marginal_seen_and_match}
\title{predict_marginal_seen_and_match}
\usage{
predict_marginal_seen_and_match(
  data = species_responses_to_model,
  model = species_attitude_model
)
}
\arguments{
\item{data}{input data for species attitude brms model object}

\item{model}{species attitude brms model object}
}
\value{
marginal predictions for seen and match
}
\description{
predict_marginal_seen_and_match
}
