% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_species_marginal_effects.R
\name{predict_species_marginal_effects}
\alias{predict_species_marginal_effects}
\title{predict_species_marginal_effects}
\usage{
predict_species_marginal_effects(data, model)
}
\arguments{
\item{data}{input data for species attitude brms model object}

\item{model}{species attitude brms model object}
}
\value{
marginal predictions for species
}
\description{
predict_species_marginal_effects
}
