% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{recolour_species_pics}
\alias{recolour_species_pics}
\title{recolour_species_pics}
\usage{
recolour_species_pics(pic, colour)
}
\arguments{
\item{pic}{jpeg object in r (array)}

\item{colour}{colour in rgb format (i.e. vector length 3)}
}
\value{
jpeg object in r (pic), recoloured to colour (colour)
}
\description{
recolour_species_pics
}
