% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_survey_data.R
\name{remove_incomplete_surveys}
\alias{remove_incomplete_surveys}
\title{remove_incomplete_surveys}
\usage{
remove_incomplete_surveys(data)
}
\arguments{
\item{data}{a data.frame containing the raw survey responses}
}
\value{
data.frame with incomplete responses removed
}
\description{
remove_incomplete_surveys
}
