% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_survey_data.R
\name{remove_instructions}
\alias{remove_instructions}
\title{remove_instructions}
\usage{
remove_instructions(data)
}
\arguments{
\item{data}{a data.frame containing the raw survey responses}
}
\value{
data.frame with the question intstructions removed
}
\description{
remove_instructions
}
