% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_metainformation.R
\name{remove_meta_info}
\alias{remove_meta_info}
\title{remove_metainformation}
\usage{
remove_meta_info(data)
}
\arguments{
\item{data}{a data.frame containing the raw survey responses}
}
\value{
data.frame with the metainformation removed
}
\description{
remove_metainformation
}
