% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_pilot_responses.R
\name{remove_pilot_responses}
\alias{remove_pilot_responses}
\title{remove_pilot_responses}
\usage{
remove_pilot_responses(data)
}
\arguments{
\item{data}{a data.frame containing the raw survey responses}
}
\value{
data.frame with responses to the pilot removed
}
\description{
remove_pilot_responses
}
