% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_response_times.R
\name{remove_response_times}
\alias{remove_response_times}
\title{remove_response_times}
\usage{
remove_response_times(data)
}
\arguments{
\item{data}{a data.frame containing the raw survey responses}
}
\value{
data.frame with the question response times removed
}
\description{
remove_response_times
}
