% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_survey_columns.R
\name{rename_final_text_Qs}
\alias{rename_final_text_Qs}
\title{rename_final_text_Qs}
\usage{
rename_final_text_Qs(data)
}
\arguments{
\item{data}{}
}
\value{

}
\description{
rename_final_text_Qs
}
