% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_mngmnt_ideas.R
\name{rename_mngmnt_ideas}
\alias{rename_mngmnt_ideas}
\title{rename_mngmnt_ideas}
\usage{
rename_mngmnt_ideas(data)
}
\arguments{
\item{data}{dataframe (respondent character table)}
}
\value{
dataframe (respondent character table with management ideas renamed)
}
\description{
rename_mngmnt_ideas
}
