% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_survey_columns.R
\name{rename_presented_species}
\alias{rename_presented_species}
\title{rename_presented_species}
\usage{
rename_presented_species(data)
}
\arguments{
\item{data}{}
}
\value{

}
\description{
rename_presented_species
}
