% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_survey_columns.R
\name{rename_recorder_info}
\alias{rename_recorder_info}
\title{rename_recorder_info}
\usage{
rename_recorder_info(data)
}
\arguments{
\item{data}{}
}
\value{

}
\description{
rename_recorder_info
}
