% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_respondent_frame.R
\name{rename_respondent_frame}
\alias{rename_respondent_frame}
\title{rename_respondent_frame}
\usage{
rename_respondent_frame(data)
}
\arguments{
\item{data}{respondent frame}
}
\value{
renamed respondent frame
}
\description{
rename_respondent_frame
}
