% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_species_of_interest.R
\name{rename_species_of_interest}
\alias{rename_species_of_interest}
\title{rename_species_of_interest}
\usage{
rename_species_of_interest(data)
}
\arguments{
\item{data}{respondent character table}
}
\value{
dataframe (respondent character table with species of interest columns renamed)
}
\description{
rename_species_of_interest
}
