% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_VoN.R
\name{rename_VoN}
\alias{rename_VoN}
\title{rename_VoN}
\usage{
rename_VoN(data)
}
\arguments{
\item{data}{dataframe (respondent character table)}
}
\value{
dataframe (respondent character table with Visions of Nature columnes renamed)
}
\description{
rename_VoN
}
