% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_supp_figures_doc.R
\name{render_supp_figures_doc}
\alias{render_supp_figures_doc}
\title{render_supp_figures_doc}
\usage{
render_supp_figures_doc(
  respondent_frame,
  species_frame,
  management_model,
  config,
  html_file_dir_out = "./doc"
)
}
\arguments{
\item{respondent_frame}{}

\item{species_frame}{}

\item{management_model}{}

\item{config}{}

\item{html_file_dir_out}{}
}
\value{
writes a word doc containing the supplementary figures
}
\description{
render_supp_figures_doc
}
