% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_survey_columns.R
\name{reorder_table_columns}
\alias{reorder_table_columns}
\title{reorder_table_columns}
\usage{
reorder_table_columns(data)
}
\arguments{
\item{data}{}
}
\value{

}
\description{
reorder_table_columns
}
