% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{scale_numerics}
\alias{scale_numerics}
\title{scale_numerics}
\usage{
scale_numerics(data)
}
\arguments{
\item{data}{respondent character table}
}
\value{
respondent character table with years recording and age mean center scaled
}
\description{
scale_numerics
}
