% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_awareness.R
\name{score_awareness}
\alias{score_awareness}
\title{score_awareness}
\usage{
score_awareness(data, numeric_out = FALSE)
}
\arguments{
\item{data}{respondent character table}

\item{numeric_out}{whether function ouput should be numeric or character (Boolean)}
}
\value{
respondent character table with awareness_score addedd either as character or numeric
}
\description{
score_awareness
}
