% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings_as_factors.R
\name{strings_as_factors}
\alias{strings_as_factors}
\title{strings_as_factors}
\usage{
strings_as_factors(data)
}
\arguments{
\item{data}{respondent character table}
}
\value{
dataframe of responses with columns converted from character to factor class
}
\description{
strings_as_factors
}
