% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_and_score_awareness.R
\name{check_and_score_awareness}
\alias{check_and_score_awareness}
\title{check_and_score_awareness}
\usage{
check_and_score_awareness(data, config)
}
\arguments{
\item{data}{respondent data}

\item{config}{config file needed for links to validation sheets}
}
\value{
respondent data with a column added for respondent's awareness of
  range-shifting species
}
\description{
check_and_score_awareness
}
