% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_awareness.R
\name{check_awareness}
\alias{check_awareness}
\title{check_awareness}
\usage{
check_awareness(data, config)
}
\arguments{
\item{data}{respondent character table}

\item{config}{config file (for paths to validated range-shifters csv)}
}
\value{
dataframe with imputed awareness added
}
\description{
check_awareness
}
