% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_education.R
\name{check_education}
\alias{check_education}
\title{check_education}
\usage{
check_education(data, config)
}
\arguments{
\item{data}{respondent character table}

\item{config}{config file (for paths to validated education csv)}
}
\value{
dataframe with imputed education added to respondent character table
}
\description{
check_education
}
