% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_respondent_character_table.R
\name{check_groups_recorded}
\alias{check_groups_recorded}
\title{check_groups_recorded}
\usage{
check_groups_recorded(data, config)
}
\arguments{
\item{data}{}

\item{config}{}
}
\value{

}
\description{
check_groups_recorded
}
