% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_respondent_character_table.R
\name{check_recorder_role}
\alias{check_recorder_role}
\title{check_recorder_role}
\usage{
check_recorder_role(data, config)
}
\arguments{
\item{data}{}

\item{config}{}
}
\value{

}
\description{
check_recorder_role
}
