% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_respondents_ages.R
\name{check_respondents_ages}
\alias{check_respondents_ages}
\title{check_respondents_ages}
\usage{
check_respondents_ages(data)
}
\arguments{
\item{data}{respondent character table}
}
\value{

}
\description{
check_respondents_ages
}
