% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_respondents_postcode.R
\name{check_respondents_postcode}
\alias{check_respondents_postcode}
\title{check_respondents_postcode}
\usage{
check_respondents_postcode(data)
}
\arguments{
\item{data}{dataframe (respondent character table)}
}
\value{
dataframe (respondent character table with postcodes renamed and formatted.
}
\description{
check_respondents_postcode
}
