% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_respondent_character_table.R
\name{check_respondents_reported_gender}
\alias{check_respondents_reported_gender}
\title{check_respondents_reported_gender}
\usage{
check_respondents_reported_gender(data)
}
\arguments{
\item{data}{}
}
\value{

}
\description{
check_respondents_reported_gender
}
