% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_species_shown_by_id.R
\name{check_species_shown_by_id}
\alias{check_species_shown_by_id}
\title{check_species_shown_by_id}
\usage{
check_species_shown_by_id(speciesdata, respondentdata)
}
\arguments{
\item{speciesdata}{cleaned species responses data}

\item{respondentdata}{cleaned  respondent data}
}
\value{
a dataframe, containing rows only for species that were in fact shown
  to respondents in the survey.
}
\description{
check_species_shown_by_id
}
