% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_wildlife_sector.R
\name{check_wildlife_sector}
\alias{check_wildlife_sector}
\title{check_wildlife_sector}
\usage{
check_wildlife_sector(data)
}
\arguments{
\item{data}{dataframe (respondent character table)}
}
\value{
dataframe (respondent character with wildlife sector column renamed
}
\description{
check_wildlife_sector
}
