% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_species_frame.R
\name{clean_species_frame}
\alias{clean_species_frame}
\title{clean_species_frame}
\usage{
clean_species_frame(data, respondent_frame)
}
\arguments{
\item{data}{species_frame}

\item{respondent_frame}{respondent_frame}
}
\value{
species_frame_cleaned
}
\description{
clean_species_frame
}
