% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_attitude_models_data.R
\name{get_attitude_models_data}
\alias{get_attitude_models_data}
\title{get_attitude_models_data}
\usage{
get_attitude_models_data(data, model = c("species", "management"))
}
\arguments{
\item{data}{survey_responses_to_model}

\item{model}{either species or management depending on which data is needed}
}
\value{
management_responses_to_model
}
\description{
get_attitude_models_data
}
