% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_from_responses_other.R
\name{impute_from_responses_other}
\alias{impute_from_responses_other}
\title{impute_from_responses_other}
\usage{
impute_from_responses_other(
  data,
  config,
  quiet = TRUE,
  view_imputations = FALSE
)
}
\arguments{
\item{data}{}

\item{config}{}

\item{quiet}{}

\item{view_imputations}{}
}
\value{

}
\description{
impute_from_responses_other
}
