% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_species_in_recorded_group.R
\name{is_species_in_recorded_group}
\alias{is_species_in_recorded_group}
\title{is_species_in_recorded_group}
\usage{
is_species_in_recorded_group(
  speciesdata,
  respondentdata = respondent_character_table_clean
)
}
\arguments{
\item{speciesdata}{cleaned species responses data}

\item{respondentdata}{cleaned respondent data}
}
\value{

}
\description{
is_species_in_recorded_group
}
