% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_attitude_pca.R
\name{make_attitude_pca}
\alias{make_attitude_pca}
\title{make_attitude_pca}
\usage{
make_attitude_pca(data)
}
\arguments{
\item{data}{respondent attitudes to management to model}
}
\value{
pca object
}
\description{
make_attitude_pca
}
