% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_management_attitudes.R
\name{model_management_attitudes}
\alias{model_management_attitudes}
\title{model_management_attitudes}
\usage{
model_management_attitudes(attitudes_to_management, respondent_data)
}
\arguments{
\item{attitudes_to_management}{management attitidues to model}

\item{respondent_data}{cleaned respondent data}
}
\value{
brms model object of predicted attitudes to management
}
\description{
model_management_attitudes
}
