% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_management_attitudes.R
\name{model_management_attitudes}
\alias{model_management_attitudes}
\title{model_management_attitudes}
\usage{
model_management_attitudes(
  species_data = management_responses_to_model,
  respondent_data = respondent_character_table_clean
)
}
\arguments{
\item{species_data}{}

\item{respondent_data}{}
}
\value{

}
\description{
model_management_attitudes
}
