% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fig_2b.R
\name{plot_fig_2b}
\alias{plot_fig_2b}
\title{plot_fig_2b}
\usage{
plot_fig_2b(data, config)
}
\arguments{
\item{data}{species effects from brms}

\item{config}{config}
}
\value{

}
\description{
plot_fig_2b
}
