% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_supp_fig_7.R
\name{plot_supp_fig_7}
\alias{plot_supp_fig_7}
\title{Title}
\usage{
plot_supp_fig_7(data = respondent_table_clean)
}
\arguments{
\item{data}{cleaned respondent data}
}
\value{
supplementary figure 7
}
\description{
Title
}
