% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_supp_fig_8_and_9.R
\name{plot_supp_fig_8}
\alias{plot_supp_fig_8}
\title{Title}
\usage{
plot_supp_fig_8(data = management_attitudes_model)
}
\arguments{
\item{data}{brms model object of a multivariate model}
}
\value{
ggplot
}
\description{
Title
}
