% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_supp_fig_8_and_9.R
\name{plot_supp_fig_9}
\alias{plot_supp_fig_9}
\title{plot_supp_fig_9}
\usage{
plot_supp_fig_9(
  data = management_responses_to_model,
  model = management_attitudes_model
)
}
\arguments{
\item{data}{Survey responses to management (model input)}

\item{model}{Management attitudes brms model}
}
\value{

}
\description{
plot_supp_fig_9
}
