% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{recolour_species_pics}
\alias{recolour_species_pics}
\title{recolour_species_pics}
\usage{
recolour_species_pics(pic, colour)
}
\arguments{
\item{pic}{}

\item{colour}{}
}
\value{

}
\description{
recolour_species_pics
}
