% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_final_text_Qs.R
\name{rename_final_text_Qs}
\alias{rename_final_text_Qs}
\title{rename_final_text_Qs}
\usage{
rename_final_text_Qs(data)
}
\arguments{
\item{data}{respondent_table_clean}
}
\value{
respondent data with a two final open text question columns renamed.
}
\description{
rename_final_text_Qs
}
