% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_presented_species.R
\name{rename_presented_species}
\alias{rename_presented_species}
\title{rename_presented_species}
\usage{
rename_presented_species(data)
}
\arguments{
\item{data}{dataframe (respondent character table)}
}
\value{
dataframe (respondent character table with species shown to respondent listed)
}
\description{
rename_presented_species
}
