% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_recorder_info.R
\name{rename_recorder_info}
\alias{rename_recorder_info}
\title{rename_recorder_info}
\usage{
rename_recorder_info(data)
}
\arguments{
\item{data}{respondent character table}
}
\value{
dataframe (respondent character table with recorder info columns renames)
}
\description{
rename_recorder_info
}
